Change update_prizetap_winning_chance_number app
